Fix when no nvidia, install pyls

+++ run.sh
@@ -10,10 +10,12 @@ source venv/bin/activate
 UNAME=$(uname)
 
 if [ "$UNAME" = "Linux" ]; then
-  OLD_PIDS="$(nvidia-smi | grep python3 | awk '{ print $5 }')"
-  if [ "$OLD_PIDS" != "" ]; then
-    echo "Killing stale python processes"
-    echo $OLD_PIDS | xargs -n1 kill -9
+  if which > nvidia-smi > /dev/null; then
+    OLD_PIDS="$(nvidia-smi | grep python3 | awk '{ print $5 }')"
+    if [ "$OLD_PIDS" != "" ]; then
+      echo "Killing stale python processes"
+      echo $OLD_PIDS | xargs -n1 kill -9
+    fi
   fi
 fi
